Conflicts:
	src/pdes/time_dependent_pde_cl.cpp

+++ src/cl_kernels/useDouble.cl
@@ -4,7 +4,7 @@
 //typedef float FLOAT;
 //typedef float4 FLOAT4;
 
-typedef float FLOAT; 
-typedef float4 FLOAT4;
+typedef double FLOAT; 
+typedef double4 FLOAT4;
 
 #endif
